pyproject.toml
[project]
name = "Valour-Laundry-Bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.6",
    "pillow",
]
